requirements.txt
Django>=4.2
djangorestframework
mysqlclient
python-dotenv
tweepy